Move two module classes to right namespace + PSR-4
These two classes were in the wrong namespace, preventing it from being
autoloaded via PSR-4. Move them to the right namespace (according to
their file system location) and add the namespace to AutoloadNamespaces.
Neither of them is used outside of Wikibase according to MediaWiki code
search [1][2], so we don’t need to bother with compatibility aliases.

This also fixes one type hint in the tests which was missing one level
of arrays.

[1]: https://codesearch.wmflabs.org/search/?q=RepoAccessModule
[2]: https://codesearch.wmflabs.org/search/?q=SitesModule

Change-Id: If0ff060ab907afd679b6fc08167a3de138371919