Move Change classes to right namespace and use PSR-4
Several classes in client/includes/Changes/ were in the wrong namespace,
preventing that directory from being autoloaded via PSR-4. Move them to
the right namespace (according to their file system location) and add
the namespace to AutoloadNamespaces.

Two of the classes are used outside of Wikibase, so a temporary
compatibility class_alias() is added for those. As in change I372453aad0
(commit 95400db8ae), MediaWiki core’s AutoLoaderStructureTest expects
the new class names to be explicitly listed in the $wgAutoloadClasses
but build/generateAutoload.php doesn’t add them there, so we briefly
have an autoload.php that’s not generated by that script, but that will
be resolved once we remove the compatibility alias.

Change-Id: Icc6c3f53e15169ea863310e8fdd7c28a09c1cacf