Validate ChangeOp against the current entity instead of an empty one
Validating on a cleared entity causes false positives like the bug
mentioned, the ChangeOp should be validated against the current state
and not on an empty entity.

This assumes removing things from an entity doesn't need validation, if
that's not true, it means "clear" would bypass that validation.

This is not ideal, a better solution would be to inject an entity differ
then diff entities, then turn that diff into a ChangeOps and apply that
on the old entity, unfortunately currently there's no class that can
turn a DiffOp to a ChangeOp, the easiest way is to write an adapter,
build the $changeRequest (imitating an API request) and then feed that
to EntityChangeOpProvider::newEntityChangeOp() but doing that would
require lots of work plus it'll be buggy and would need lots of
refinements to be ready for use.

Depends-On: If95c32244170226d9549fd11d2ba741039c76e34
Bug: T243158
Change-Id: I9de4b9202a9ab7923078ac263722a572228b12df