bridge: ensure base styles in storybook
f15304e restricted to which parts of storybook our application styles
should apply, and included the reset styles as part of them to ensure
future consistent use.

This patch ensures that the "base styles" we want to apply in storybook,
but which are not existing in the application code itself, do not get
reset by the production reset styles.

An alternative to the weird strong selector could be to use a
storybook-specific selector (e.g. #storybook-preview-iframe) and
configure it in prefixwrap to be ignored[0] - but that would have us
touch/augment production config for a dev-only need.

[0] https://www.npmjs.com/package/postcss-prefixwrap#ignored-selectors

Change-Id: I4c807bbc877a4bcf5dc776f3504a0f5411190a1e