bridge: fix cascade-protected messages
The messages for cascade protection on the repo and client both used a
$2 parameter for the project’s administrators page, which was not
documented and not provided by the code. Add it to ErrorPermission.

The client version of the message actually doesn’t need a parameter – it
can just form the link using normal wikilink syntax, just like the link
to the project’s cascade page. The English version of the message is
updated accordingly, and the parameter is not documented in qqq.json –
however, since a few translations already use $2, we still provide the
parameter in the code for now. Hopefully those translations will be
updated soon, and then we can drop the parameter from the client version
from the message (though this will complicate the test somewhat).

The repo version of the message does need the parameter, since the link
needs to be formed using the repo router. The parameter could be $1,
since it’s the only one – but existing translations already use $2, and
that number is also consistent with the “administrators” link in the
“protected on repo” message, so we keep it and mark $1 as unused.

Bug: T243918
Change-Id: Iaf8cb3ba3ba53c046bb87488196c81c5090cb517