wbterms: Fix incorrect deletion of rows in findActuallyUnusedTermIds
This is essentially the exact same problem as in
I4db2069a781e392a7de9d7a606ef795a7b3cae19.
During that investigation and commit, Tom and I looked only at
DatabaseTermIdsCleaner, not at this extra level that was in the
TermStore itself (since they moved to a trait for reuse).

We will do some followups fixing more names in var and also
trying to write some more tests for this bit of code too...

Peer programmed with @Tarrow

Bug: T243944
Bug: T243705
Change-Id: I519d2daba95d08fbb9a9855290419a238c73f8f0