bridge: report error on deprecated statements
Deprecated statements should not be edited via the Wikidata Bridge, at
least not for the moment. Check the rank when validating bridge
applicability and error out if it’s “deprecated”.

To implement this, a new PathToStatement interface is added, analogous
to the existing PathToSnak, used to navigate from a full entity state to
a single statement. MainSnakPath, for simplicity’s sake, implements
both, and most of the former resolveSnakInStatement() code now makes up
resolveStatement(). Maybe PathToStatement can also be used to access
references in the future.

Bug: T243888
Change-Id: Ib066a56ff565d1a38954c355216710a225aa4301