bridge: fix storybook for vuex-smart-module
This is basically a missing part of change I0e449857c8 (commit
7638b6f2f8): not just the tests need to use the real module, but the
stories which use the store as well.

I couldn’t find a way to restore the original rootModule.options.state
(there’s no afterEach() equivalent in storybook decorators), but as long
as each story which uses the store overrides the state, that shouldn’t
be a problem.

Change-Id: If8cca61a4b9e85f581b9dc039952964a31a32b7a