Add missing wbeditentity summary messages for clients
For summaries formatted in Wikibase Repos these kinds of message can be
prefixed with wikibase-[ENTITY_TYPE] where the entity type is derived
from the title associated with the edit. This does not happen on
Wikibase clients, so the message has to have an equivalent prefixed
with `wikibase-entity-`.

Compare `RepoHooks::onFormat` and `ClientHooks::onFormat`.

Bug: T244129
Change-Id: If0491492b4824e2e71bfcc4db2510ebdf8553eae