bridge: fix horizontal spacing of ErrorUnsupportedDatatype
Remove the default padding-left on IconMessageBox. I don’t think it
makes sense to have any padding or margin on an inline IconMessageBox
(the non-inline version already overrides all four padding directions at
once), and it was causing problems in ErrorUnsupportedDatatype (the only
place where IconMessageBox is used so far), because that should have the
same space on both sides.

Instead, move that space (as margin rather than padding, matching other
components) to the ErrorUnsupportedDatatype component, via a class on
the whole IconMessageBox. The BailoutActions also get a class, because
it seems to me that we might want to pull its margin into the parent
components as well – but that won’t happen in this change.

Bug: T235753
Change-Id: Ibcd48c7d5bfa8e590d73d63bef0786417c80438c