bridge: remove webdriverio from package.json
This is just confusing – it doesn’t make a difference which version is
specified here, the effective version is whatever the @wdio/* packages
pull in, and those usually pin exact versions. And as far as I
understand, specifying a top-level webdriverio dependency has no effect
when it matches the @wdio/* packages, and when it doesn’t match, it
prevents those packages’ versions from being deduplicated being being
pulled into the top-level node_modules/, so it actually has a
detrimental effect.

Change-Id: I32943e0152b68a44b8a19facb8b3b5b5ce568041