bridge: Upgrade vue dependencies
This updates @vue/cli-service to fix a moderate security vulnerability
as reported by `npm audit`.

This comes with number of other major updates which makes some migration
changes necessary.

The babel config was adjust automagically by vue upgrade.

The jest config was adjust to use new presets according to
https://cli.vuejs.org/migrating-from-v3/#vue-cli-plugin-unit-jest
ts-jest is no longer needed.

vue-eslint-parser was added manually. It is unclear why it is not
installed to node_modules automatically as it is a dependency of
eslint-plugin-vue.

`jest.mock()` now allows only access to a very restricted set of
variables or those prefixed with `mock. Thus prefixing that variable
with `mock`.

Change-Id: I942d3be3533ab4d01206ebc5a6950268d65d84d1