bridge: saner font size handling
This should make the handling of sizes easier to understand and be
consistent on desktop and mobile.

The font-size: 1rem in App.vue ensures that we are basing our sizes
directly on the font-size of the browser. Especially, this is intended
to negate the effects of the font-size: 0.875em on
body > .oo-ui-windowManager on the vector skin.

All relative sizes are now rem, that means only relative to the root
font-size, not relative to the parent font-size. This is done in order
to be able to reason about the actual size of a value defined in
_variables.scss, without the actual size of $base-spacing-unit changing
depending on the font-size of the current or parent element.

Consequently, it removes the font-size from the global storybook scss as
that is no longer affecting us in production.

Some of the sizes and names in this commit are likely to be changed in
future commits when syncing with UX.

It also aligns some font-sizes to the current UX design, which mandates
that headings are 1rem in size.

Bug: T243192
Change-Id: I4bc8cf4962c82b8db0ea9d976fb49edfbaec9d20