bridge: remove store {action,mutation,getter}Types
I think these were just cargo culting from the pre-vuex-smart-module
days – with vuex-smart-module, the definitions and uses are just as
type-safe without dedicated constants for these names. (The only thing
that’s not checked by TypeScript, I believe, is the use in test names,
as in “it dispatches $someName”.)

It is unclear whether using getter types was ever reasonable.

Change-Id: Ic69a2dbe8895fa00efec42b837a30eddc089923d