bridge: show property labels in bailout components
The label is extracted into a component to make it easy to use in the
other components (since it has to be passed into the message plugin, it
can’t be a slot or something like that).

The stories of the components using the TermLabel are updated to add a
property label and also add the property ID to the URL where it was
missing. Marie Curie’s item didn’t have suitable statements for all
cases, so I looked for some other items using the query service.

The targetLabel getter checks whether the label is known by comparing it
to null; if it’s undefined, the getter thinks the label is defined and
returns it, causing errors elsewhere. To avoid this, make the test
store’s state class extend the real state class, ensuring that the
members are initialized with their default values.

To test the components which now inject HTML elements into the message
plugin, the calledWithHTMLElement helper introduced in change
I973e3a7695 (commit cb78b5716c) is extracted into a dedicated module.

Bug: T235753
Bug: T238660
Bug: T240212
Bug: T242747
Change-Id: Ia06cffe73471046e3783105fe989bb81741b90bc