Refactor entity data updates from Content to Handler
This makes use of some of the new core features that were
introduced as part of MCR.
All DataUpdate stuff is now in the Handlers instead of in the
Content.
I took this oppourtunity to also remove the logic from
the shared Entity class and instead put everything specifically
in ItemHandler and PropertyHandler.

This change also removed 2 Wikibase provided PHP hooks:
 - WikibaseEntityModificationUpdate
 - WikibaseEntityDeletionUpdate
I did not find any usage of these hooks on Gerrit or Github
with all references to that string only refering to forks of
Wikibase and the initial hook definition.

The term store writer is now also pushed down into the
2 handlers rather than being in the generic handler which is
also used by entities that don't care about secondary term
storage.

Change-Id: Ifcf8c20292302615aad3b630ad32802c9406a22c