bridge: fix vertical marging of bailout heading
This was affected by the same problem that was fixed for the button in
change Ibde86e5eb6 (commit f8408c380c) – the margin is in em, the
heading had a larger font-size, so the margin grew unexpectedly. Scale
the difference away for now, though we might want to move away from the
px-to-em() function soon (see change I4bc8cf4962).

Bug: T235753
Bug: T238660
Bug: T240212
Bug: T242747
Change-Id: If0bd830d90584ada398ba3da0d54cc2957079a38