wbterms: Split new term storage cleaning into own transaction
This change keeps the general logic of the new terms store the same
in terms of writes, however the cleanup is moved to a job and thus
also a seperate transaction.
I considered just splitting the deffered update into 2 transactions,
but really there is no need for the cleanup to be done in a deffered
update.

Cleanup could probably still be split into 3 transactions (one per
table). As with the previous commit in this chain this will decrease the
likelyhood that we encounter locks and also for WMF production decrease
load on main app servers.

As the cleanup is in its own transaction all of the writes that are part
of T244115 get committed earlier (pre cleanup) for other requests to
see.
This also means that we might be able to SELECT FOR UPDATE rows during
insertions (as the transaction will be much shorter) without issues.
And just generally lock things in more places that we kind of want to.

Bug: T244115
Change-Id: I480749aff7761d95c763e40d15bfdba9234516e5