bridge: add targetValue to root state
Instead of having a targetValue getter which gets the value from the
statements module, actually save the targetValue in the root state. For
now, it is still *also* updated in the statements module, but we intend
to change that next, so that the statements module is only updated after
a successful save.

Bug: T238662
Change-Id: Idffe0a61731180a1ca6b7c7a469ff832b3b94da7