bridge: define vertical margins in DataBridge-component
These vertical margins between component are conceptually important and
depend on the order of the components, not so much on the component
itself. Therefore it makes sense to create them in the component that
actually does the ordering.

This patch is the alternative to a different patch that implements this
with a spacing component instead of classes.

Bug: T239421
Change-Id: I94bec97fd48f18719d154bc5b61dad24936481cb
Co-authored-by: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>