Move more lib classes to the correct namespace
To keep this commit simple, it only moves classes that are *not* used in
any code outside of this repository, i. e. the ones where we don’t have
to bother with compatibility class aliases. The remaining ones will be
moved afterwards.

I’ve tried to keep imports roughly sorted lexicographically, though
they’re not always sorted to begin with.

Change-Id: I66674ba13a2eb48f6791a2d818ba4cfc4856f590