bridge: put targetValue into statement only on save
This moves the mutation of the statementState to only after successful
persistence (and round trip) to the server, allowing us to attempt a
retry in the future.
Consequently, the originalStatement state became obsolete and could be
removed.
Detection of changed state (by user input) happens by comparison between
targetValue and the contents of statementModule.
applyStringDataValue would not need to be an actual store action (maybe
it is even somewhat confusing) and it can move out when implementing the
different patch strategies.

One downside of state update after successful save only: As actions are
not tracked in vue dev tools[0], this makes for poor visibility of what
is going on - the "mutations" are performed only on a copy of the
statements state before submitting the change, which is not visible in
the otherwise very helpful browser extension.

I'm unsure about the value of the applyStringDataValue integration tests
as they don't integrate a lot.

[0] https://github.com/vuejs/vue-devtools/issues/179

Bug: T238662
Change-Id: Id0a9f656ee06f15fb814e6ecd724b934c50bf6ad