bridge: Upgrade @types/jest to match actual jest version
This brings our jest types in line with the version that is actually
already installed.
Also, this change doesn't seem to be fully mature yet, as it doesn't
recognize that the type 'repo' is satisfied by the value 'repo', which
would lead to non-sensical things like `'repo' as 'repo'`. In these
cases I disabled type checking for that line.

Change-Id: Icdd6e0909949da225ef1685b2b61f26010ed6157