bridge: Update stylelint and disable unit blacklist
This updates both stylelint and the wikimedia configuration which we
are extending.

The wikimedia-stylelint-configuration has since 0.8 a new rule that uses
a large cannon to kill a small bird and is not sensical in our use-case
so it is deactivated. See wikimedia/stylelint-config-wikimedia#94
While it is true that the unit rem can be used wrongly in subtle ways
and there are contexts where em is the better choice overall, e.g.
the vuejs-components-library, the benefit of a much lower cognitive
overhead in an integrated app is not insignificant either.

Change-Id: I57da32bd9e0cbfd1e099beed0d0852f4790248ac