bridge: isolate statement state from outside
Persist _a copy_ of the argument into the statement state to avoid
unintentional modification of it through references to the original
tree (or vice versa).
This was a problem in a test but would be just as undesireable in prod,
or even more so.
Placement of the clone can probably be debated but the mutation is a
central spot which makes for a great candidate.

C.f. https://gerrit.wikimedia.org/r/plugins/gitiles/wikibase/termbox/+/0d4af42/src/common/EntityInitializer.ts#8

Change-Id: I7d05c0f42c31633332fe6e512bb5a8004b869e29