bridge: Align ul styling
This provides basic styling for ul to all components. In particular, the
goal is to align the indentation in the bailout actions component to
that in the page list of the cascading error component.

Components themselves can then implement deviations from it, e.g.
different spacing between list items in the bailout actions, or no
list-style in the reference section.

Bug: T239421
Change-Id: Ic76b1f8bb3fa7b3f31661d55feb48dd60eb6433e