bridge: Centralize styling for <strong>, <b>, <em> and <i>
A few components and messages use these tags and their styling is defined
multiple times accross the codebase.
This patch centralizes them in _reset.css (inspired by Ic76b1f8bb)

It also moves the _reset.scss to almost the end of the imports in
main.scss as the files that do not define styles, e.g. variables and
mixins, should come before those that do. This allows the reset css to
access variables in variables.scss.

Bug: T239421
Change-Id: I92b8fed8df88a4bff243aa96d103c3209bdb423f