bridge: show intent in e2e assertion
Before this change the API payload was asserted correctly, but only
"conincidentally". Both application state and test (testSet) referred
to the same object and the application mutated both by reference
(with the result of uuid) - making the significance of the assert
questionable at best.
This came to light when introducing isolation of the statement state
(later patch).

Also makes the interface of the SpecialPageReadingEntityRepository
response available to the test.

Change-Id: I0d0e949c8ea46fc76ceb1a0e567bd94285255402