bridge: prepare for different statement mutations
Allow for different mutation strategies for statements and move the
(only) existing implemenation into the ReplaceMutationStrategy.

We could not agree on a limit of how far reaching the changes
performed by these strategies could reasonably be, so the
whole of StatementState is passed, offering endless possibilities.

I considered cloning targetValue before passing it into the mutation
strategy to avoid it being mutated "by accident", but we have vuex warn
us about that automatically so we should be good.

Bug: T238662
Change-Id: I9bf15cabee3a23d3350e3a323666573f33384f49