bridge: fix EventEmittingButton sizes
This separates the EventEmittingButton into two sizes (see Figma [1]),
“L” and “M”. The “L” size is similar to the previous behavior (though
the font-size of 16px is now explicitly specified), but with more
padding that had previously been missing; the “M” size has a smaller
font size and less padding. To ensure the intended heights of 40px and
32px respectively, the line-height is now defined directly on the
component. (Previously, it already got the correct line-height from the
app, but not in the storybook.)

[1]: https://www.figma.com/file/KIkcQ7NjOl8OkQtGWe5duIbW/Wikidata-Bridge?node-id=2840%3A2

Bug: T239421
Change-Id: I4a8ecaea24bd6f2b4002d9e3879fb25511ff7e5f