Use quasi-open metadata key rather than branded one
The corporate platform seems to fallback correctly, see
https://developer.twitter.com/en/docs/tweets/optimize-with-cards/overview/markup

But we should reconsider the "summary" value, in this case.  Using the
card-compatible type name is nice for [corporate platform], but I suspect that
we could use the value "website" and still see a summary card.

This shouldn't change any behavior, it's just a "no logo" cleanup.

Bug: T157145
Change-Id: I361213dd20adede33aeb94b42230fe452ade6d48