bridge: drastically simplify MainSnakPath
I noticed while looking into some other changes that this implementation
can be simplified a lot with the new optional chaining [1] and nullish
coalescing [2] operators in TypeScript 3.7, which we use since change
I0d7f6228276 (commit 916421f47a).

This requires a small change to the ts-loader options in the storybook
webpack config, otherwise the ?. syntax won’t be transformed.

[1]: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#optional-chaining
[2]: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#nullish-coalescing

Change-Id: I68fe8dc91810f7b97632a6ea6938febc81eb7868