bridge: hint only relevant statement content in logic
Interestingly, we can use the EntityId hint in Record - we have been
struggeling with something comparable earlier
(Id6fcb39e898c94b57cda100be28a274fff43fac9).

The brevity and lack of good place to put it makes me shy away from
glorifying this in its own file. It's not data model, and not a
data-access definition either (maybe by shape, but not intent). It's
somewhat bound to the store, but then we'd come full circle.

The name 'state' in MutationStrategy may still be controversial but it's
nice and short and entitiesStatementMaps or EntityStatementMapMap don't
quite have a ring to them.

Change-Id: I3e326607a69507beb32a8bed199d298d407322ae