Do not reference mediawiki.legacy.shared in view/resources.php "for snakview"
The addition of `mediawiki.legacy.shared` is  a "skin-level" decision
NOT an extension-level decision as loading it can lead to loading incompatible
or redundant code.

The module is loaded on all skins except Minerva.

In preparation for the deprecation of mediawiki.legacy.shared I propose removing this
module here to avoid log warnings.

For Minerva is this change breaks any display of "snakview" I suggest copying and
pasting the appropriate code into a skinStyle for the module that only targets Minerva
to make this code more future proof.

Bug: T242177
Change-Id: I661210d778ee1f12bd63c7a0674f16f96720cc28