bridge: guard getters against initializing lifecycle
Change the getters to what their guards, I believe, actually meant to
say - providing a consumer-friendly mock result when the application is
still in its initializing phase (that's what their tests asserted all
along).

I'm not sure this is the best approach generally but should overcome the
defect described in T244999#5914867. Will think a bit more about if
these checks are truly necessary and if/how we can do them in a better
way.

Bug: T244999
Change-Id: Iea762c8b5c00138ce9c3988bc822e552f4851f70