bridge: allow interaction with Loading in storybook
This introduces @storybook/addon-knobs and makes use of it to showcase
the behavior of the Loading component.
As a sideeffect it resolves warnings due to missing required properties
(for Loading in its story).

Also updates storybook and friends to latest patch version.

See https://www.npmjs.com/package/@storybook/addon-knobs

Bug: T237433
Change-Id: I81fa886a0d336db444472584e2baa6c7d7e432a8