bridge: declare Statement ID as optional
In a statement we receive from the API, the ID is always present.
However, when we create new statements (for the “update” edit decision),
we want to leave the ID blank so that the server will assign a fresh one
(avoiding the need to implement this, including knowledge of what
statement IDs look like, in Bridge). The easiest way to do this, for
now, is to reuse the Statement interface and make its ID member
optional. (We may eventually need to go for an alternative approach,
where there are two interfaces for statements with optional and required
IDs, and corresponding changes to StatementMap etc. – but not yet.)

And while we’re at it, also add comments to the other optional members
of the interface.

Bug: T238662
Change-Id: I99c240499b66a60ac826b39eb7ee80743df002ce