bridge: make input field larger on mobile
The specification from UX is that it should have a height of 40px,
including border; the height is in fact variable depending on the
content (it’s a ResizingTextField), but 7px of vertical padding produce
a height of 40px for a single line of input.

Note that the ResizingTextField will not resize itself on window resize,
so if the window is resized to be smaller or larger than the breakpoint,
then the height will be incorrect until the next time the input changes.

Bug: T239421
Change-Id: I05f9fea7112988c9b97b05d361ba4d7f9c5792b4