bridge: Add StatementGroup getter to MainSnakPath
This refactoring completes the list of resolvers on MainSnakPath for every
statement-related step.

It also generalizes the some getters that can in principle work for
references and qualifiers as well to use the new method.

This allows us to make the properties on MainSnakPath private and
enforce access through the methods defined by the interfaces.

Also, since both the current and the foreseeable future implementation
of StatementMutationStrategy work for all kinds of StatementGroups, we
can have the interface depend on other interfaces instead of
MainSnakPath directly.

Change-Id: Iabfee32cdc97924223e22ff90247e444a82a3b20