bridge: extract marginForCenterColumn mixin
This generalizes the former marginInputComponent mixin into one that’s
used for any component on the “center column”, with 24px margin to the
left and right, but no margin on mobile. The mixin has some extra
conditionals to make the emitted CSS minimal – the margin-bottom is left
out if it’s identical to the margin-top, and both are 0 (common),
simplify the mobile case even further. The end result is that the
generated CSS doesn’t change at all.

Some changes to the stylelint config are necessary to format the
@if/@else nicely (by default, stylelint wants a blank line before the
@else, which looks silly). The at-rule-empty-line-before config is that
from stylelint-config-wikimedia, except for the ignoreAtRules, which is
the newly needed part.

Change-Id: I797663d74491ab08af9e94ec6211776f1e38d9df