bridge: Add UpdateMutationStrategy
The guards are not yet extracted as only the "datavaluetype is the same"
seems to potentially be the same across all strategies and I'm not sure
that this merits the overhead of a base class.

The actual usage in the action will be implemented in a later commit as
it seems to be sufficiently disconnected by having the EditDecision
value sufficiently hard-coded.

Bug: T238662
Change-Id: Idbd7b90e973710a7b5976010568a4b852fcd891a