bridge: make EventEmittingButton sizes responsive
The button in the bailout actions should be a size “M” by default, but a
size “L” on mobile. Since this is currently our only size “M” button in
the app, and it seems reasonable to assume that future uses of the
button should also be responsive, the “M” size is made responsive by
default, being equivalent to “L” on mobile. (Though a later change will
make “L” equivalent to a new size “XL” on mobile.) To avoid duplication,
the actual size of the button is extracted into a local mixin, which is
then included depending on the size class and @media query.

To test this in Storybook, the “open canvas in new tab” button (top
right, middle of three buttons) is very useful, to avoid the rest of the
storybook UI distorting the sizes.

Bug: T239421
Change-Id: I258f322c129fe32570a9b58972b33ddb94caec26