bridge: make header larger on mobile
The header height variable is updated to more closely match the
(non-iconless) EventEmittingButton size: 38px (mobile: 46px) are
relative to the browser font size (px-to-rem), 2px (mobile: still 2px)
are absolute (px). The cancel button size is, as before, tied directly
to the header size, which isn’t very nice, but good enough for now.

Bug: T239421
Change-Id: I77463d687678fe774f483183d59a961ecf35b6c5