Do not hold DB connections in class properties in new term store classes
A pattern emerged in the new term storage related db classes that
maintains DB connections as class properties.
This goes against current patterns, and means that a reference to the db
connection may hang around for longer than expected and this may not be
desirable.

Use of the pattern removed in this commit already lead to some other
undesired patterns trying to creep into code, such as creating db
connection handles in service constructors.

Lets remove the pattern to avoid temptation.

Bug: T246468
Change-Id: I1d5a2012138f84e435b665c2fb149692948688fb