bridge: disable value debounce for knobs in storybook
So far we have no use for debounced values from knobs in storybook - we
only use few knobs and the updates seem to be quite performant.
We filed storybookjs/storybook#10019 [0] to check with the storybook
maintainers if the behavior we witnessed is as expected.
While researching we saw that work [1] was put into this area recently
so maybe more fine grained controls will be possible with the next
release, which we then can use if need arises.

[0] https://github.com/storybookjs/storybook/issues/10019
[1] https://github.com/storybookjs/storybook/pull/9447

Change-Id: I3ea30acb60e142dff5c9fbf4fe4cda9a40d4c990