bridge: separate Loading stories more clearly
The “initializing” and “saving” props can’t both be true simultaneously,
so it’s not very useful to have separate “initializing” and “saving”
stories where both props can be set each time. Instead, make each story
only responsible for one prop, and hard-code the other to false.

With that, getLoadingProps() can be simplified to remove the options
parameter, and hard-code the default timings (we never specified other
values anyways). And while we’re at it, also use friendlier labels for
them – the storybook isn’t just for developers, after all.

(Note that, while getLoadingProps() always returns “the same” object, it
still has to be a function and the two stories still need to call
number() separately – it can’t be extracted into a constant.)

Change-Id: If5a1e4c8c1066ca7a5303dd518b48d36840d8f3c