bridge: remove repeated assertion in test
I assume that this assertion was initially introduced in one test and
then copy+pasted a bunch of times (and I’m almost certainly one of the
guilty parties). This doesn’t seem very useful to me – if we want to
assert that it returns a ServiceContainer, let’s do that in a separate
test, and let the other tests concentrate on asserting their actually
significant conditions.

Change-Id: I790fdc7810d24c3fb5a46cad114383ec20e71783