bridge: Add license settings to config API module
Using the copyrightpage message is what Termbox seems to be using. It
can be assumed that these terms of use are only changing very rarely and
will be kept in sync with other places that define the terms of use,
e.g. WikimediaMessages.

The callback is injected because the Title class is still a 4500+ lines
monster class and not exactly easy to use and test. That should be
refactored after the Title class has been refactored.

Bug: T246225
Change-Id: I3e5a86804358c72cedc8dfb8710413d324c773ad