bridge: add size M for icon-only EventEmittingButton
Like ordinary size-M buttons, it’s 32px high (14px font-size × 1.5
line-height + 4px padding-top + 5px padding-bottom + 2×1px border). The
size of the icon itself is unchanged – the visible part is 14×14px, as
specified.

The button also grows to size L on mobile, consistent with other button
sizes (ordinary and icon-only); in practice, this probably won’t matter,
as Figma specifies that on mobile, the license banner (the only place
where we use this size variant of the button) grows to occupy the whole
area below the header, and the ordinary “cancel” button is replaced with
a “back” button, making the button inside the banner obsolete.

Bug: T246227
Change-Id: I204689305386046c1045942ee6ec42915bb34d5f