Allow injection of HttpRequestFactory in services
This now allows injection of a HttpRequestFactory
but does not allow calling of any of the methods.

It would be nice to factor this out into a trait that could
be shared across repo and client, but I'm going to leave that
to a future patch, which would also create somewhere nice for
such things to live.

Change-Id: I2b4d6affbdcfde599d25420a414078a309ee029e