Fix buggy logic on building change op result of aliases
The old and new aliases sent to buildResult() are the actual old and new
aliases of the entity, not the old and new aliases sent to apply()
meaning all of this complex logic is unnecessary and actually buggy
(see the bug).

The only complicating factor here is changing order of aliases.
Just changing order of aliases currently cause a valid edit with empty diff.
In order to comply with Wikibase, this considers them an actual edit. We
can reconsider this later.

Added some more tests.

Bug: T246873
Change-Id: Ie53f624b901f4d0ee9990250f69a0dcada60aac9