Do not try to load terms from cache in getPrefetched*()
The TermBuffer interface requires that the prefetched Terms are not
loaded from the cache but from an instance attribute. However, the
previous implementation only guaranteed that for terms that are not
already cached.  That is a breach of the contract given by the
TermBuffer interface that does not allow for prefetched terms to be not
available if they exist.  This patch mends that broken contract by
prefetching the data from the cache as well.

It uses a local buffer for the Terms retrieved from the cache, but
accesses the terms retrieved from the database via the inner prefetching
lookup in order to not duplicate them and conserve memory.

Checking for valid language codes was moved to the margin to ensure that
all class internal calls always use valid languages.

Bug: T247196
Change-Id: Ic63f2607ed1d543ff9f7ba18db6f479c59f9841b
Co-authored-by: Michael Große <mic.grosse@googlemail.com>
Co-authored-by: Tobias Andersson <tobias.andersson@wikimedia.de>