bridge: Simplify BridgeConfig plugin
This plugin gets options that are enforced to be type-safe thanks to
typescript. No need to do test that here again. Also that definition of
the config options type is redundant and not used anywhere else.

We are initialising that plugin with only the client options and then
later create it anew with client and repo options. So we need to account
for that.

Change-Id: I1a02b290aeaf7d8be98a78ff385323fd9c1462cc