bridge: add border-top for header on mobile
On desktop, the whole App is surrounded by a border from the OOUI
window, so we only define a border-bottom for the ProcessDialogHeader,
but not a border-top. But on mobile, the OOUI window occupies the whole
window and no longer has a border, so in that case we need to define our
own border-top if the whole header should still be shown with a border
on both sides and be 50px high in total.

Bug: T239421
Change-Id: Ia0e16be2e22d089c86aa2f5c9f11a49ffea7ee9b