PSR-4 autoload as much as possible of repo
AutoloadClasses takes precedence over AutoloadNamespaces, and
build/generateAutoload.php can recognize which classes are correctly
namespaced (can use AutoloadNamespaces) and which are not (need
AutoloadClasses), so we can already declare the fully general autoload
namespace in the extension-repo-wip.json file, and let the autoload
generator override those classes that still need it in AutoloadClasses.

This also turns the repo/autoload.php file into a convenient work list:
every class in there must be moved to the correct namespace.

This is essentially a resubmit of change Id3d8cac3e3, which had done the
same thing for lib, repo, and client a while ago; at the time, MediaWiki
core’s AutoLoaderStructureTest had rejected the overly general
AutoloadNamespaces, but with some recent changes they are now allowed.
(Lib and client were already made PSR-4 compliant in the meantime, so
this resubmit now only affects repo.)

Bug: T172368
Change-Id: Ic0e095985d917c7693c08ecbbf14d5e5be8440e5
Depends-On: Ie5e50ecb519b99a1197688c046c7be245ce6da1b