bridge: share a Tracker instance between init and app
Emit tracking events to the same underlying Tracker instance in
both the init and app part of bridge.

Also introduces a PrefixingEventTracker which decorates Trackers,
prefixing the emitted topic - to avoid redundantly specifying those
prefixes. It is neither init nor app-specific so it found a home in a
concern-specific module folder. The interface should probably go there
as well (I don't fancy dedicated folders for interfaces only).

Creating the topic strings through logic decreases their discoverability
(e.g. in full-text search), but it provides other benefits. The
application-specific part of the string can, at least for now, still be
found where first configured. Source of the e2e tests may be a last
resort for full strings but some are dynamic by nature (e.g. datatype).

This addresses a comment from back in
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Wikibase/+/551188/10/client/data-bridge/src/mediawiki/init.ts#41
and the topic of redundancy came up again in
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Wikibase/+/578324/4/client/data-bridge/src/data-access/DataBridgeTrackerService.ts#5

Change-Id: Ibb00b8d62dca091f37e22d6d97e050f5de37a587